Small modifications to testDriver to make it read a file and create output suitable for ssa-search ( > to file ).
SSA-Search takes in the output of the testDriver and looks for a query.

Few issues:
is it possible to make ensure that the suffix array contructor only
includes whole multiples of the mask length
	example: 
act ctc ggc at! @#$ % >> 101 >> ac cc gc a! @$

reason why I ask is that it's a heck of a lot harder to realign the
suffixes to the original without knowing where the masking stopped--not
impossible, just more work. You could also include the position from
which each suffix is derived, which would help me avoid having to thrawl
through T' for each suffix in the match range.